frontend: Reorganise Thirty Year Regulation file structure
- Rename main component to `ThirtyYearRegulationTab`
- Remove unused exports
- Combine test component files
