+++ frontend/src/features/functions/ThirtyYearRegulationTabComponents/ThirtyYearResultListItem.tsx
