+++ frontend/src/features/functions/ThirtyYearRegulationTabComponents/index.ts
@@ -1,8 +1,7 @@
 export {default as ExternalSalesDataImport} from "./ExternalSalesDataImport";
 export {default as ThirtyYearErrorModal} from "./ThirtyYearErrorModal";
+export {default as ThirtyYearErrorTest} from "./ThirtyYearErrorTest";
 export {default as ThirtyYearLoadedResults} from "./ThirtyYearLoadedResults";
 export {default as ThirtyYearResultListItem} from "./ThirtyYearResultListItem";
 export {default as ThirtyYearResults} from "./ThirtyYearResults";
 export {default as ThirtyYearSkippedList} from "./ThirtyYearSkippedList";
-export {default as ThirtyYearErrorTest} from "./ThirtyYearTestComponents/ThirtyYearErrorTest";
-export {default as ThirtyYearErrorTestResult} from "./ThirtyYearTestComponents/ThirtyYearErrorTestResult";
