+++ frontend/src/features/functions/index.ts
@@ -1,3 +1,3 @@
 export {default as FunctionsPage} from "./FunctionsPage";
 export {default as SurfaceAreaPriceCeilingTab} from "./SurfaceAreaPriceCeilingTab";
-export {default as ThirtyYearRegulation} from "./ThirtyYearRegulation";
+export {default as ThirtyYearRegulation} from "./ThirtyYearRegulationTab";
